Изменил сущности и конфигурации, поменял первую миграцию
