+++ DataAccess/Extensions/DbRegisterExtension.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace DataAccess.Extensions
+{
+    public static class DbRegisterExtension
+    {
+        public static IServiceCollection AddDbContext(this IServiceCollection serviceCollection,
+        IConfiguration configuration)
+        {
+            return serviceCollection.AddDbContext<AppDbContext>(builder =>
+            {
+                builder.UseNpgsql(configuration["Database:ConnectionString"]);
+                builder.UseSnakeCaseNamingConvention();
+            });
+        }
+    }
+}
