+++ Domain/Entities/MovieContent.cs
@@ -3,5 +3,6 @@
 public class MovieContent : ContentBase
 {
     public long MovieLength { get; set; }
+    public string VideoUrl { get; set; } = null!;
     public DateOnly ReleaseDate { get; set; }
 }
