+++ Domain/Entities/PersonInContent.cs
@@ -5,7 +5,7 @@ public class PersonInContent
     public int Id { get; set; }
     
     public long ContentId { get; set; }
-    public ContentBase Content { get; set; }
+    public ContentBase Content { get; set; } = null!;
 
     public string Name { get; set; } = null!;
