+++ Domain/Entities/SeasonInfo.cs
@@ -4,8 +4,9 @@ public class SeasonInfo
 {
     public int Id { get; set; }
 
-    public SerialContent SerialContent { get; set; } = null!;
+	public int SeasonNumber { get; set; }
 
-    public int SeasonNumber { get; set; }
-    public int EpisodesCount { get; set; }
+	public SerialContent SerialContent { get; set; } = null!;
+
+    public List<Episode> Episodes { get; set; } = null!;
 }
